Add global query modal action
