+++ libs/features/query/src/QueryResults/QueryResults.tsx
@@ -43,12 +43,15 @@ import {
   useConfirmation,
 } from '@jetstream/ui';
 import {
+  QueryHistory,
+  QueryHistoryRef,
   ViewEditCloneRecord,
   fromJetstreamEvents,
   fromQueryHistoryState,
   fromQueryState,
   isAsyncJob,
   useAmplitude,
+  useQueryRestore,
 } from '@jetstream/ui-core';
 import { getFlattenSubqueryFlattenedFieldMap } from '@jetstream/ui-core/shared';
 import { fromAppState, googleDriveAccessState } from '@jetstream/ui/app-state';
@@ -60,9 +63,7 @@ import React, { Fragment, useCallback, useEffect, useRef, useState } from 'react
 import { Link, NavLink, useLocation, useNavigate } from 'react-router-dom';
 import { useRecoilState, useRecoilValue } from 'recoil';
 import { filter } from 'rxjs/operators';
-import QueryHistory, { QueryHistoryRef } from '../QueryHistory/QueryHistory';
 import IncludeDeletedRecordsToggle from '../QueryOptions/IncludeDeletedRecords';
-import useQueryRestore from '../utils/useQueryRestore';
 import QueryResultsAttachmentDownload, { FILE_DOWNLOAD_FIELD_MAP } from './QueryResultsAttachmentDownload';
 import QueryResultsCopyToClipboard from './QueryResultsCopyToClipboard';
 import QueryResultsDownloadButton from './QueryResultsDownloadButton';
