+++ libs/shared/ui-core-shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from './query-fields-utils';
+export * from './query-filter.utils';
 export * from './query-soql-utils';
 export * from './query-utils';
