+++ libs/shared/ui-core/src/index.ts
@@ -61,6 +61,10 @@ export * from './orgs/OrgSelectionRequired';
 export * from './orgs/OrgWelcomeInstructions';
 export * from './orgs/useOrgPermissions';
 export * from './orgs/useUpdateOrgs';
+export * from './query/QueryHistory/QueryHistory';
+export * from './query/RestoreQuery/RestoreQuery';
+export * from './query/RestoreQuery/useQueryRestore';
+export * from './query/ValidQueryText';
 export * from './record/RecordSearchPopover';
 export * from './record/ViewChildRecords';
 export * from './record/ViewEditCloneRecord';
