+++ libs/shared/ui-core/src/query/QueryHistory/QueryHistoryItemCard.tsx
@@ -5,7 +5,6 @@ import { APP_ROUTES } from '@jetstream/shared/ui-router';
 import { pluralizeFromNumber } from '@jetstream/shared/utils';
 import { QueryHistoryItem, SalesforceOrgUi } from '@jetstream/types';
 import { ButtonGroupContainer, Card, CopyToClipboard, Grid, GridCol, Icon, Spinner, Textarea } from '@jetstream/ui';
-import { useAmplitude } from '@jetstream/ui-core';
 import { queryHistoryDb } from '@jetstream/ui/db';
 import { isQueryValid } from '@jetstreamapp/soql-parser-js';
 import Editor from '@monaco-editor/react';
@@ -13,8 +12,9 @@ import { formatDate } from 'date-fns/format';
 import clamp from 'lodash/clamp';
 import { FunctionComponent, useEffect, useRef, useState } from 'react';
 import { Link } from 'react-router-dom';
-import RestoreQuery from '../QueryBuilder/RestoreQuery';
-import { SoqlValidIndicator } from '../utils/ValidQueryText';
+import { useAmplitude } from '../..';
+import { RestoreQuery } from '../RestoreQuery/RestoreQuery';
+import { SoqlValidIndicator } from '../ValidQueryText';
 
 const SOBJECT_QUERY_ICON: IconObj = { type: 'standard', icon: 'record_lookup', description: 'Object Query' };
 const METADATA_QUERY_ICON: IconObj = { type: 'standard', icon: 'settings', description: 'Metadata Query' };
