+++ libs/shared/ui-core/src/query/ValidQueryText.tsx
